requirements.txt
Flask
Gunicorn
flask-cors
openai
python-dotenv
elevenlabs
numpy
pydub
pyTelegramBotAPI